requirements-pip.txt


Kivy==1.11.1
pygame==1.9.6
requests==2.23.0
Pillow==7.1.2